SHIT UI NOW WORKS, FUCK THIS IS AMAZING!!!
so it loops and works, motion is visible.
a bit more UI will be added to impl different view/conf renderers, etc, upscaing crap

+++ MotionScope/GpuWorker/AsyncGpuWorker.h
@@ -13,14 +13,14 @@ namespace gpu {
 		using cuda::motion::render::ViewType;
 
 		struct Job {
-			size_t frameIndex;
+			int frameIndex;
 			size_t generation;
 			QImage prev;
 			QImage curr;
 		};
 
 		struct Result {
-			size_t frameIndex;
+			int frameIndex;
 			size_t generation;
 			QImage prev;
 			QImage curr;
